+++ packages/happner-cluster/RELEASES.md
@@ -346,9 +346,10 @@
   - TEN-103: set default timeout on component loading wait warning to 30secs
   - TEN-104: errors are only logged on system failures
 
-12.0.7 2022-03-16
------------------ 
-  - SMC-4817: updated happn-cluster
+12.1.0 2022-04-08
+-----------------
+  - TEN-95: $call discovers components and methods in cluster
+